Tickets #185, #220,
git-svn-id: https://svn.code.sf.net/p/tcnopen/trdp/trunk@1779 3b5a3598-5f4e-4449-9e63-bd40438bfec0

+++ trdp/src/common/trdp_utils.c
@@ -179,9 +179,9 @@ static BOOL8 trdp_SockDelJoin (
  *   Globals
  */
 
-UINT32 trdp_getCurrentMaxSocketCnt()
+INT32 trdp_getCurrentMaxSocketCnt()
 {
-    return (UINT32)sCurrentMaxSocketCnt;
+    return sCurrentMaxSocketCnt;
 }
 
 /**********************************************************************************************************************/
